Fixes floating point errors in calculating final result

+++ quiz/views.py
@@ -114,7 +114,7 @@ def get_student_response(request):
                     wrong_answers += 1
                     iterator += 1
 
-            marks_obtained = float(correct_answers - wrong_answers * negative_marks)
+            marks_obtained = float(correct_answers - float(wrong_answers * negative_marks))
             total_questions = sum(questions_in_each_section)
             total_correct_answers = sum(correct_ans_each_section)
             total_wrong_answers = sum(wrong_in_each_section)
